+++ packages/backend-deployer/src/cdk_deployer.test.ts
@@ -8,6 +8,7 @@ import {
 import { DeployProps } from './cdk_deployer_singleton_factory.js';
 import { CdkErrorMapper } from './cdk_error_mapper.js';
 import { UniqueBackendIdentifier } from '@aws-amplify/plugin-types';
+import { BackendDeployerEnvironmentVariables } from './environment_variables.js';
 
 void describe('invokeCDKCommand', () => {
   const uniqueBackendIdentifier: UniqueBackendIdentifier =
@@ -34,7 +35,7 @@ void describe('invokeCDKCommand', () => {
   void it('handles no options/args', async () => {
     await invoker.deploy();
     assert.strictEqual(execaMock.mock.callCount(), 1);
-    assert.equal(execaMock.mock.calls[0].arguments[1]?.length, 6);
+    assert.equal(execaMock.mock.calls[0].arguments[1]?.length, 8);
     assert.deepStrictEqual(execaMock.mock.calls[0].arguments[1], [
       'cdk',
       'deploy',
@@ -42,13 +43,15 @@ void describe('invokeCDKCommand', () => {
       '--app',
       "'npx tsx amplify/backend.ts'",
       '--all',
+      '--output',
+      '.amplify/artifacts/cdk.out',
     ]);
   });
 
   void it('handles options for branch deployments', async () => {
     await invoker.deploy(uniqueBackendIdentifier);
     assert.strictEqual(execaMock.mock.callCount(), 1);
-    assert.equal(execaMock.mock.calls[0].arguments[1]?.length, 10);
+    assert.equal(execaMock.mock.calls[0].arguments[1]?.length, 12);
     assert.deepStrictEqual(execaMock.mock.calls[0].arguments[1], [
       'cdk',
       'deploy',
@@ -56,6 +59,8 @@ void describe('invokeCDKCommand', () => {
       '--app',
       "'npx tsx amplify/backend.ts'",
       '--all',
+      '--output',
+      '.amplify/artifacts/cdk.out',
       '--context',
       'backend-id=123',
       '--context',
@@ -66,7 +71,7 @@ void describe('invokeCDKCommand', () => {
   void it('handles deployProps for sandbox', async () => {
     await invoker.deploy(undefined, sandboxDeployProps);
     assert.strictEqual(execaMock.mock.callCount(), 1);
-    assert.equal(execaMock.mock.calls[0].arguments[1]?.length, 12);
+    assert.equal(execaMock.mock.calls[0].arguments[1]?.length, 14);
     assert.deepStrictEqual(execaMock.mock.calls[0].arguments[1], [
       'cdk',
       'deploy',
@@ -74,6 +79,8 @@ void describe('invokeCDKCommand', () => {
       '--app',
       "'npx tsx amplify/backend.ts'",
       '--all',
+      '--output',
+      '.amplify/artifacts/cdk.out',
       '--context',
       'deployment-type=SANDBOX',
       '--hotswap-fallback',
@@ -88,7 +95,7 @@ void describe('invokeCDKCommand', () => {
   void it('handles options and deployProps for sandbox', async () => {
     await invoker.deploy(uniqueBackendIdentifier, sandboxDeployProps);
     assert.strictEqual(execaMock.mock.callCount(), 1);
-    assert.equal(execaMock.mock.calls[0].arguments[1]?.length, 14);
+    assert.equal(execaMock.mock.calls[0].arguments[1]?.length, 16);
     assert.deepStrictEqual(execaMock.mock.calls[0].arguments[1], [
       'cdk',
       'deploy',
@@ -96,6 +103,8 @@ void describe('invokeCDKCommand', () => {
       '--app',
       "'npx tsx amplify/backend.ts'",
       '--all',
+      '--output',
+      '.amplify/artifacts/cdk.out',
       '--context',
       'backend-id=123',
       '--context',
@@ -114,7 +123,7 @@ void describe('invokeCDKCommand', () => {
       deploymentType: BackendDeploymentType.SANDBOX,
     });
     assert.strictEqual(execaMock.mock.callCount(), 1);
-    assert.equal(execaMock.mock.calls[0].arguments[1]?.length, 11);
+    assert.equal(execaMock.mock.calls[0].arguments[1]?.length, 13);
     assert.deepStrictEqual(execaMock.mock.calls[0].arguments[1], [
       'cdk',
       'destroy',
@@ -122,6 +131,8 @@ void describe('invokeCDKCommand', () => {
       '--app',
       "'npx tsx amplify/backend.ts'",
       '--all',
+      '--output',
+      '.amplify/artifacts/cdk.out',
       '--context',
       'backend-id=123',
       '--context',
@@ -136,14 +147,20 @@ void describe('invokeCDKCommand', () => {
       validateAppSources: true,
     });
     assert.strictEqual(execaMock.mock.callCount(), 2);
-    assert.equal(execaMock.mock.calls[0].arguments[1]?.length, 4);
+    assert.equal(execaMock.mock.calls[0].arguments[1]?.length, 10);
     assert.deepStrictEqual(execaMock.mock.calls[0].arguments[1], [
       'tsc',
       '--noEmit',
       '--skipLibCheck',
+      '--module',
+      'node16',
+      '--moduleResolution',
+      'node16',
+      '--target',
+      'es2022',
       'amplify/backend.ts',
     ]);
-    assert.equal(execaMock.mock.calls[1].arguments[1]?.length, 12);
+    assert.equal(execaMock.mock.calls[1].arguments[1]?.length, 14);
     assert.deepStrictEqual(execaMock.mock.calls[1].arguments[1], [
       'cdk',
       'deploy',
@@ -151,6 +168,8 @@ void describe('invokeCDKCommand', () => {
       '--app',
       "'npx tsx amplify/backend.ts'",
       '--all',
+      '--output',
+      '.amplify/artifacts/cdk.out',
       '--context',
       'backend-id=123',
       '--context',
@@ -166,14 +185,20 @@ void describe('invokeCDKCommand', () => {
       validateAppSources: true,
     });
     assert.strictEqual(execaMock.mock.callCount(), 2);
-    assert.equal(execaMock.mock.calls[0].arguments[1]?.length, 4);
+    assert.equal(execaMock.mock.calls[0].arguments[1]?.length, 10);
     assert.deepStrictEqual(execaMock.mock.calls[0].arguments[1], [
       'tsc',
       '--noEmit',
       '--skipLibCheck',
+      '--module',
+      'node16',
+      '--moduleResolution',
+      'node16',
+      '--target',
+      'es2022',
       'amplify/backend.ts',
     ]);
-    assert.equal(execaMock.mock.calls[1].arguments[1]?.length, 10);
+    assert.equal(execaMock.mock.calls[1].arguments[1]?.length, 12);
     assert.deepStrictEqual(execaMock.mock.calls[1].arguments[1], [
       'cdk',
       'deploy',
@@ -181,11 +206,81 @@ void describe('invokeCDKCommand', () => {
       '--app',
       "'npx tsx amplify/backend.ts'",
       '--all',
+      '--output',
+      '.amplify/artifacts/cdk.out',
       '--context',
       'deployment-type=SANDBOX',
       '--hotswap-fallback',
       '--method=direct',
     ]);
+  });
+
+  void it('overrides enabled type checking for branch deployments', async () => {
+    try {
+      process.env[
+        BackendDeployerEnvironmentVariables.ALWAYS_DISABLE_APP_SOURCES_VALIDATION
+      ] = 'true';
+      await invoker.deploy(uniqueBackendIdentifier, {
+        deploymentType: BackendDeploymentType.BRANCH,
+        validateAppSources: true,
+      });
+      assert.strictEqual(execaMock.mock.callCount(), 1);
+      assert.equal(execaMock.mock.calls[0].arguments[1]?.length, 14);
+      assert.deepStrictEqual(execaMock.mock.calls[0].arguments[1], [
+        'cdk',
+        'deploy',
+        '--ci',
+        '--app',
+        "'npx tsx amplify/backend.ts'",
+        '--all',
+        '--output',
+        '.amplify/artifacts/cdk.out',
+        '--context',
+        'backend-id=123',
+        '--context',
+        'branch-name=testBranch',
+        '--context',
+        'deployment-type=BRANCH',
+      ]);
+    } finally {
+      delete process.env[
+        BackendDeployerEnvironmentVariables
+          .ALWAYS_DISABLE_APP_SOURCES_VALIDATION
+      ];
+    }
+  });
+
+  void it('overrides enabled type checking for sandbox deployments', async () => {
+    try {
+      process.env[
+        BackendDeployerEnvironmentVariables.ALWAYS_DISABLE_APP_SOURCES_VALIDATION
+      ] = 'true';
+      await invoker.deploy(undefined, {
+        deploymentType: BackendDeploymentType.SANDBOX,
+        validateAppSources: true,
+      });
+      assert.strictEqual(execaMock.mock.callCount(), 1);
+      assert.equal(execaMock.mock.calls[0].arguments[1]?.length, 12);
+      assert.deepStrictEqual(execaMock.mock.calls[0].arguments[1], [
+        'cdk',
+        'deploy',
+        '--ci',
+        '--app',
+        "'npx tsx amplify/backend.ts'",
+        '--all',
+        '--output',
+        '.amplify/artifacts/cdk.out',
+        '--context',
+        'deployment-type=SANDBOX',
+        '--hotswap-fallback',
+        '--method=direct',
+      ]);
+    } finally {
+      delete process.env[
+        BackendDeployerEnvironmentVariables
+          .ALWAYS_DISABLE_APP_SOURCES_VALIDATION
+      ];
+    }
   });
 
   void it('returns human readable errors', async () => {